Removed (already existing elsewhere) openstreetmap link generator on res.partner and renamed module into website_snippet_openstreetmap
