--- setup/website_openstreetmap/odoo/addons/website_openstreetmap
@@ -1,1 +1,0 @@
-../../../../website_openstreetmap
