+++ setup/website_snippet_openstreetmap/odoo/addons/website_snippet_openstreetmap
@@ -1,0 +1,1 @@
+../../../../website_snippet_openstreetmap
