+++ setup/website_snippet_openstreetmap/setup.py
