--- website_openstreetmap/README.rst
@@ -1,5 +1,0 @@
-=====================
-Website Openstreetmap
-=====================
-
-Website Openstreetmap
