--- website_openstreetmap/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_website_openstreetmap
