+++ website_snippet_openstreetmap/README.rst
@@ -1,0 +1,5 @@
+=============================
+Website Snippet OpenstreetMap
+=============================
+
+Website Snippet OpenstreetMap
