+++ website_snippet_openstreetmap/__manifest__.py
@@ -1,8 +1,8 @@
 # Copyright 2023 Onestein (<http://www.onestein.eu>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
-    "name": "Website Open Street Map",
-    "summary": "Website Open Street Map",
+    "name": "Website Snippet OpenStreetMap",
+    "summary": "Website Snippet OpenStreetMap",
     "category": "Website",
     "version": "15.0.1.0.0",
     "author": "Onestein, Odoo Community Association (OCA)",
@@ -19,10 +19,10 @@
     "assets": {
         "website.assets_wysiwyg": [
             ("include", "web._assets_helpers"),
-            "website_openstreetmap/static/src/snippets/s_openstreetmap/options.js",
+            "website_snippet_openstreetmap/static/src/snippets/s_openstreetmap/options.js",
         ],
         "web.assets_frontend": [
-            "website_openstreetmap/static/src/js/utils.js",
+            "website_snippet_openstreetmap/static/src/js/utils.js",
         ],
     },
 }
