+++ website_snippet_openstreetmap/readme/CONFIGURE.rst
