+++ website_snippet_openstreetmap/readme/CONTRIBUTORS.rst
