+++ website_snippet_openstreetmap/readme/DESCRIPTION.rst
