+++ website_snippet_openstreetmap/readme/USAGE.rst
