+++ website_snippet_openstreetmap/views/snippets/snippets.xml
@@ -7,8 +7,8 @@
             position="after"
         >
             <t
-                t-snippet="website_openstreetmap.s_openstreetmap"
-                t-thumbnail="/website_openstreetmap/static/src/img/snippets_thumbs/s_openstreetmap.svg"
+                t-snippet="website_snippet_openstreetmap.s_openstreetmap"
+                t-thumbnail="/website_snippet_openstreetmap/static/src/img/snippets_thumbs/s_openstreetmap.svg"
             />
         </xpath>
     </template>
